[metrics]: Allow EPP to register metrics from extention (kubernetes-sigs/gateway-api-inference-extension#1787)
The motivation is
https://github.com/llm-d/llm-d-inference-scheduler/issues/386 that we
would like to have more metrics from llm-d inference scheduler plugin.

The current metrics implementation is bundled in the runner of EPP, so
this PR extends the runner to allow extention to register prometheus
metrics collector.

See https://github.com/llm-d/llm-d-inference-scheduler/pull/405 for a
draft PR of how metrics are going to be implemented in the plugin.

I have tested locally that both EPP and inference scheduler metrics can
be exported through the pod /metrics endpoint.